Nullify _length on reset or filter change
_length nullified on filter change only if length is 0

This causes length to be re-fetched automatically if switching away from a filter with 0 results. This is desirable in all use cases, including scenarios where the length is returned from the server with every request. (If length is 0, objectAt won't initiate any requests for data.)

If length is fetched in a distinctive request, manually call the .requestLength method on the sparse array from your controller/component/route/etc.